examples/13-a2a-gateway/requirements.txt
# Runtime dependencies for example 13 — A2A Gateway
# The soorma and soorma-common packages are installed from the workspace (see start.sh).
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
